+++ requirements-testing.txt
@@ -1,4 +1,5 @@
 aioresponses >= 0.6.0
+cryptography >= 2.7.0
 docker >= 3.1.1
 flake8 >= 3.5.0
 pytest-cov >= 2.5.1
